refactor(agent): extract createGraph function to separate module
- Move createGraph function from deepModeling.ts to createGraph.ts
- Update imports in deepModeling.ts to use extracted function
- Improve code organization and separation of concerns for workflow setup
