+++ frontend/internal-packages/agent/src/deepModeling.ts
@@ -1,29 +1,15 @@
 import { AIMessage, HumanMessage } from '@langchain/core/messages'
 import { RunCollectorCallbackHandler } from '@langchain/core/tracers/run_collector'
-import { END, START, StateGraph } from '@langchain/langgraph'
 import type { Schema } from '@liam-hq/db-structure'
 import type { Result } from 'neverthrow'
 import { err, ok } from 'neverthrow'
 import { v4 as uuidv4 } from 'uuid'
 import { WORKFLOW_ERROR_MESSAGES } from './chat/workflow/constants'
-import {
-  analyzeRequirementsNode,
-  designSchemaNode,
-  executeDdlNode,
-  finalizeArtifactsNode,
-  generateUsecaseNode,
-  prepareDmlNode,
-  validateSchemaNode,
-  webSearchNode,
-} from './chat/workflow/nodes'
-import {
-  createAnnotations,
-  DEFAULT_RECURSION_LIMIT,
-} from './chat/workflow/shared/langGraphUtils'
+import { finalizeArtifactsNode } from './chat/workflow/nodes'
+import { DEFAULT_RECURSION_LIMIT } from './chat/workflow/shared/langGraphUtils'
 import type { WorkflowConfigurable, WorkflowState } from './chat/workflow/types'
 import { withTimelineItemSync } from './chat/workflow/utils/withTimelineItemSync'
-import { invokeSchemaDesignToolNode } from './db-agent/nodes/invokeSchemaDesignToolNode'
-import { routeAfterDesignSchema } from './db-agent/routing/routeAfterDesignSchema'
+import { createGraph } from './createGraph'
 
 export type DeepModelingParams = {
   userInput: string
@@ -38,98 +24,6 @@ export type DeepModelingParams = {
 }
 
 export type DeepModelingResult = Result<WorkflowState, Error>
-
-/**
- * Retry policy configuration for all nodes
- */
-const RETRY_POLICY = {
-  maxAttempts: process.env['NODE_ENV'] === 'test' ? 1 : 3,
-}
-
-/**
- * Create and configure the LangGraph workflow
- */
-const createGraph = () => {
-  const ChatStateAnnotation = createAnnotations()
-  const graph = new StateGraph(ChatStateAnnotation)
-
-  graph
-    .addNode('webSearch', webSearchNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('analyzeRequirements', analyzeRequirementsNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('designSchema', designSchemaNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('invokeSchemaDesignTool', invokeSchemaDesignToolNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('executeDDL', executeDdlNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('generateUsecase', generateUsecaseNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('prepareDML', prepareDmlNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('validateSchema', validateSchemaNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-    .addNode('finalizeArtifacts', finalizeArtifactsNode, {
-      retryPolicy: RETRY_POLICY,
-    })
-
-    .addEdge(START, 'webSearch')
-    .addEdge('webSearch', 'analyzeRequirements')
-    .addEdge('analyzeRequirements', 'designSchema')
-    .addEdge('invokeSchemaDesignTool', 'designSchema')
-    .addConditionalEdges('designSchema', routeAfterDesignSchema, {
-      invokeSchemaDesignTool: 'invokeSchemaDesignTool',
-      executeDDL: 'executeDDL',
-    })
-    .addEdge('executeDDL', 'generateUsecase')
-    .addEdge('generateUsecase', 'prepareDML')
-    .addEdge('prepareDML', 'validateSchema')
-    .addEdge('finalizeArtifacts', END)
-
-    // Conditional edge for executeDDL - retry with designSchema if DDL execution fails
-    .addConditionalEdges(
-      'executeDDL',
-      (state) => {
-        if (state.shouldRetryWithDesignSchema) {
-          return 'designSchema'
-        }
-        if (state.ddlExecutionFailed) {
-          return 'finalizeArtifacts'
-        }
-        return 'generateUsecase'
-      },
-      {
-        designSchema: 'designSchema',
-        finalizeArtifacts: 'finalizeArtifacts',
-        generateUsecase: 'generateUsecase',
-      },
-    )
-
-    // Conditional edges for validation results
-    .addConditionalEdges(
-      'validateSchema',
-      (state) => {
-        // success → finalizeArtifacts
-        // dml error or test fail → designSchema
-        return state.error ? 'designSchema' : 'finalizeArtifacts'
-      },
-      {
-        designSchema: 'designSchema',
-        finalizeArtifacts: 'finalizeArtifacts',
-      },
-    )
-
-  return graph.compile()
-}
 
 /**
  * Execute Deep Modeling workflow
